llm-customenv/requirements.txt
azureml-core==1.52.0
azureml-dataset-runtime==1.52.0
azureml-defaults==1.52.0
azure-ml==0.0.1
azure-ml-component==0.9.18.post2
azureml-mlflow==1.52.0
azureml-contrib-services==1.52.0
azureml-contrib-services==1.52.0
azureml-automl-common-tools==1.52.0
torch-tb-profiler~=0.4.0
azureml-inference-server-http~=0.8.0
inference-schema~=1.5.0
MarkupSafe==2.1.2
regex
pybind11
bitsandbytes==0.39.1
git+https://github.com/huggingface/transformers.git
git+https://github.com/huggingface/peft.git
git+https://github.com/huggingface/accelerate.git
datasets
scipy
azure-identity